feat: Add RTCM v3 legacy messages 1004-1033, 1042, 1046, MSM3/5/7, and comprehensive tests

+++ src/Asv.Gnss/Protocols/Rtcm/V3/RtcmV3Message.cs
@@ -26,7 +26,7 @@ public abstract class RtcmV3MessageBase : IProtocolMessage<ushort>
         }
 
         bitIndex += 6; // reserved
-        var messageLength = (byte)SpanBitHelper.GetBitU(buffer,ref bitIndex, 10);
+        var messageLength = (int)SpanBitHelper.GetBitU(buffer,ref bitIndex, 10);
         if (messageLength > (buffer.Length - 3 /* preamble-8bit + reserved-6bit + length-10bit  */ - 3 /* crc 24 bit */))
         {
             throw new Exception($"Deserialization RTCMv3 message failed: length too small. Want '{messageLength}'. Read = '{buffer.Length - 6}'");
